+++ pytorch_hccl_tests/osu/collectives/__init__.py
@@ -1,5 +1,6 @@
-from pytorch_hccl_tests.osu.collectives.osu_allreduce import allreduce  # noqa
-from pytorch_hccl_tests.osu.collectives.osu_broadcast import broadcast  # noqa
-from pytorch_hccl_tests.osu.collectives.osu_alltoall import alltoall  # noqa
 from pytorch_hccl_tests.osu.collectives.osu_allgather import allgather  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_allreduce import allreduce  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_alltoall import alltoall  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_barrier import barrier  # noqa
+from pytorch_hccl_tests.osu.collectives.osu_broadcast import broadcast  # noqa
 from pytorch_hccl_tests.osu.collectives.osu_reduce_scatter import reducescatter  # noqa