+++ pytorch_hccl_tests/osu/collectives/osu_allgather.py
@@ -1,10 +1,15 @@
-from time import perf_counter_ns as now
 import logging
 import pandas as pd
 
 import torch.distributed as dist
 
-from pytorch_hccl_tests.commons import get_device, safe_rand
+from pytorch_hccl_tests.commons import (
+    get_device,
+    safe_rand,
+    get_device_event,
+    sync_device,
+    elaspsed_time_ms,
+)
 from pytorch_hccl_tests.osu.options import Options
 from pytorch_hccl_tests.osu.osu_util_mpi import Utils
 
@@ -39,13 +44,15 @@ def allgather(args):
         dist.barrier()
         for i in iterations:
             if i == options.skip:
-                tic = now()
+                start_event = get_device_event(backend)
             dist.all_gather(tensor_list, tensor, pg, False)
-        toc = now()
+        end_event = get_device_event(backend)
+        sync_device(backend)
         dist.barrier()
 
+        total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
         avg_latency = Utils.avg_lat(
-            toc, tic, 2 * options.iterations, world_size, device
+            total_time_ms, 2 * options.iterations, world_size, device
         )
 
         if rank == 0: