+++ pytorch_hccl_tests/osu/collectives/osu_alltoall.py
@@ -1,11 +1,16 @@
 import logging
-from time import perf_counter_ns as now
 
 import pandas as pd
 import torch
 import torch.distributed as dist
 
-from pytorch_hccl_tests.commons import get_device, get_dtype
+from pytorch_hccl_tests.commons import (
+    elaspsed_time_ms,
+    get_device,
+    get_device_event,
+    get_dtype,
+    sync_device,
+)
 from pytorch_hccl_tests.osu.options import Options
 from pytorch_hccl_tests.osu.osu_util_mpi import Utils
 
@@ -39,13 +44,15 @@ def alltoall(args):
         dist.barrier()
         for i in iterations:
             if i == options.skip:
-                tic = now()
+                start_event = get_device_event(backend)
             dist.all_to_all_single(out_tensor, in_tensor)
-        toc = now()
+        end_event = get_device_event(backend)
+        sync_device(backend)
         dist.barrier()
 
+        total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
         avg_latency = Utils.avg_lat(
-            toc, tic, 2 * options.iterations, world_size, device
+            total_time_ms, 2 * options.iterations, world_size, device
         )
 
         if rank == 0: