+++ pytorch_hccl_tests/osu/collectives/osu_broadcast.py
@@ -1,10 +1,15 @@
 import logging
-from time import perf_counter_ns as now
 
 import pandas as pd
 import torch.distributed as dist
 
-from pytorch_hccl_tests.commons import get_device, safe_rand
+from pytorch_hccl_tests.commons import (
+    elaspsed_time_ms,
+    get_device,
+    get_device_event,
+    safe_rand,
+    sync_device,
+)
 from pytorch_hccl_tests.osu.options import Options
 from pytorch_hccl_tests.osu.osu_util_mpi import Utils
 
@@ -38,13 +43,15 @@ def broadcast(args):
         dist.barrier()
         for i in iterations:
             if i == options.skip:
-                tic = now()
+                start_event = get_device_event(backend)
             dist.broadcast(msg, 0, pg, False)
-        toc = now()
+        end_event = get_device_event(backend)
+        sync_device(backend)
         dist.barrier()
 
+        total_time_ms = elaspsed_time_ms(backend, start_event, end_event)
         avg_latency = Utils.avg_lat(
-            toc, tic, 2 * options.iterations, world_size, device
+            total_time_ms, 2 * options.iterations, world_size, device
         )
 
         if rank == 0:
@@ -54,6 +61,5 @@ def broadcast(args):
                 ignore_index=True,
             )
 
-    # Persist result to CSV file
     if rank == 0:
         df.to_csv(f"osu_broadcast-{device.type}-{dtype}-{world_size}.csv", index=False)