+++ pytorch_hccl_tests/osu/p2p/osu_multi_lat.py
@@ -55,8 +55,9 @@ def multi_lat(args):
                 dist.send(s_msg, partner, pg, 1)
             toc = now()
 
+        elapsed_time_ms = (toc - tic) / 1000
         avg_latency = Utils.avg_lat(
-            toc, tic, 2 * options.iterations, world_size, device
+            elapsed_time_ms, 2 * options.iterations, world_size, device
         )
 
         if rank == 0: